pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crier"
version = "0.4.0"
description = "Cross-post your content to dev.to, Hashnode, Medium, and more"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Alex Towell", email = "lex@metafunctor.com" }
]
keywords = ["cross-posting", "devto", "hashnode", "medium", "blogging", "content"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet",
    "Topic :: Utilities",
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "pyyaml>=6.0",
    "rich>=13.0",
    "questionary>=2.0",
    "pyperclip>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
crier = "crier.cli:cli"

[project.urls]
Homepage = "https://github.com/queelius/crier"
Repository = "https://github.com/queelius/crier"
Issues = "https://github.com/queelius/crier/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/crier"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
